requirements.txt
streamlit
pandas
python-dotenv
tavily-python

# LangChain 패키지 (가상 환경용 최종 검증 버전)
langchain==0.2.6
langchain-core==0.2.10
langchain-community==0.2.6
langgraph==0.1.4
langchain-openai==0.1.9
langchain-experimental==0.0.62